% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_change.R
\name{delta_change}
\alias{delta_change}
\title{Compute delta change}
\usage{
delta_change(db_object, start = 0, end = 6, method = "none")
}
\arguments{
\item{db_object}{tables in a human exercise object generated by fetchdata}

\item{start}{starting interval for delta change}

\item{end}{ending interval for delta change}

\item{method}{ending interval for delta change}
}
\description{
Fetches data from in-memory / local sqlite database
}
\examples{
\dontrun{
hello()
}
}
