% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchdata.R
\docType{class}
\name{human_exercise-class}
\alias{human_exercise-class}
\title{fetchdata}
\arguments{
\item{db}{name of local database}

\item{experiments}{vector of queried experiments}

\item{fields}{vector of queried fields}

\item{omit_subjects}{vector of subjects to omit from query}

\item{selecte_interval}{vector of intervals to query}

\item{exact_match}{exact match?}

\item{include_age}{include age}
}
\description{
Fetches data from in-memory / local sqlite database
}
\examples{
\dontrun{
hello()
}
}
