% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grids.R
\docType{data}
\name{hump_colours}
\alias{hump_colours}
\title{this runs the grid plots}
\format{An object of class \code{character} of length 2.}
\usage{
hump_colours
}
\arguments{
\item{xvector}{choose experiment for the x axis}

\item{yvector}{choose experiment for the y axis}

\item{type}{choose experiment for the y axis}
}
\description{
Runs a bunch of correlations and then picks the best ones
}
\examples{
\dontrun{
hello()
}
}
\keyword{datasets}
