% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/john_the_fisherman.R
\name{john_the_fisherman}
\alias{john_the_fisherman}
\title{Compute delta change}
\usage{
john_the_fisherman(experiment1, experiment2, filter_exercise = NULL,
  stat = "P", hit_limit = 15, plot = "matrix", selected_x = NULL,
  selected_y = NULL, xinterval = 0, yinterval = 0,
  split.gender = F)
}
\arguments{
\item{experiment1}{choose experiment for the x axis}

\item{experiment2}{choose experiment for the y axis}

\item{filter_exercise}{filter by group. If null, uses all.}

\item{stat}{choose statistic r/p.value to sort the results by}

\item{hit_limit}{choose number of results to print out / plot}

\item{plot}{choose how to plot e.g. matrix for printouts or plot for plots}

\item{selected_x}{choose how to plot e.g. matrix for printouts or plot for plots}
}
\description{
Runs a bunch of correlations and then picks the best ones
}
\examples{
\dontrun{
hello()
}
}
