% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grids.R
\name{plot_group}
\alias{plot_group}
\title{this runs the grid plots}
\usage{
plot_group(ex, df, xvar, yvar, xlimit = c(0, 110), ylimit = c(0, 50),
  annotx)
}
\arguments{
\item{xvector}{choose experiment for the x axis}

\item{yvector}{choose experiment for the y axis}

\item{type}{choose experiment for the y axis}
}
\description{
Runs a bunch of correlations and then picks the best ones
}
\examples{
\dontrun{
hello()
}
}
