% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accelerometry.R
\name{process_accelerometry}
\alias{process_accelerometry}
\title{process_accelerometry}
\usage{
process_accelerometry(Subject, nci.methods = F,
  start.date = as.Date("2014/1/5"), start.time = "00:00:00",
  id = NULL, brevity = 2, valid.days = 1, valid.week.days = 0,
  valid.weekend.days = 0, int.axis = "vert", int.cuts = c(100, 760,
  1100, 1500), cpm.nci = FALSE, hourly.axis = "vert",
  days.distinct = FALSE, nonwear.axis = "vert", nonwear.window = 90,
  nonwear.tol = 0, nonwear.tol.upper = 99, nonwear.nci = FALSE,
  weartime.minimum = 600, weartime.maximum = 1440,
  partialday.minimum = 1440, active.bout.length = 10,
  active.bout.tol = 0, mvpa.bout.tol.lower = 0,
  vig.bout.tol.lower = 0, active.bout.nci = F, sed.bout.tol = 0,
  artifact.axis = "vert", artifact.thresh = 25000,
  artifact.action = 1, weekday.weekend = F, return.form = 2)
}
\arguments{
\item{n}{Number of files}
}
\description{
Processes accelerometry data from .agd files
}
\examples{
\dontrun{
hello()
}
}
