% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accelerometry.R
\name{process_accelerometry}
\alias{process_accelerometry}
\title{process_accelerometry}
\usage{
process_accelerometry(df)
}
\arguments{
\item{n}{Number of files}
}
\description{
Processes accelerometry data from .agd files
}
\examples{
\dontrun{
hello()
}
}
